+++ src/Text/Gigaparsec/Debug.hs
@@ -15,11 +15,11 @@ breakpoints that can be used to pause parsing execution.
 module Text.Gigaparsec.Debug (debug, debugWith, debugConfig, DebugConfig(..), WatchedReg(..), Break(..)) where
 
 import Text.Gigaparsec.Internal (Parsec)
-import Text.Gigaparsec.Internal.RT (Reg, readReg)
 import Text.Gigaparsec.Internal qualified as Internal
-import Text.Gigaparsec.Internal.RT qualified as Internal
 
+import Data.Ref (Ref, readRef)
 import Control.Monad (when, forM)
+import Control.Monad.RT.Unsafe (RT, unsafeIOToRT)
 import System.IO (hGetEcho, hSetEcho, hPutStr, stdin, stdout, Handle)
 import Data.List (intercalate, isPrefixOf)
 import Data.List.NonEmpty (NonEmpty((:|)), (<|))
@@ -54,7 +54,7 @@ tracked, which is why this datatype is existential.
 -}
 type WatchedReg :: *
 data WatchedReg = forall r a. Show a => WatchedReg String    -- ^ the name of the register
-                                                   (Reg r a) -- ^ the register itself
+                                                   (Ref r a) -- ^ the register itself
 
 {-|
 This is used by 'DebugConfig' to specify whether the parsing should be paused
@@ -94,7 +94,7 @@ on either entry, exit, or both. The parse is resumed by entering any character o
 debugWith :: DebugConfig -> String -> Parsec a -> Parsec a
 debugWith config@DebugConfig{ascii} name (Internal.Parsec p) = Internal.Parsec $ \st good bad -> do
   -- TODO: could make it so the postamble can print input information from the entry?
-  ascii' <- (\colourful -> ascii || not colourful) <$> Internal.unsafeIOToRT supportsPretty
+  ascii' <- (\colourful -> ascii || not colourful) <$> unsafeIOToRT supportsPretty
   let config' = config { ascii = ascii' }
   doDebug name Enter st ""  config'
   let good' x st' = do
@@ -127,12 +127,12 @@ shouldBreak :: Direction -> Break -> Bool
 shouldBreak Enter = breakOnEntry
 shouldBreak Exit = breakOnExit
 
-doDebug :: String -> Direction -> Internal.State -> String -> DebugConfig -> Internal.RT ()
+doDebug :: String -> Direction -> Internal.State -> String -> DebugConfig -> RT ()
 doDebug name dir st end DebugConfig{..} = do
   printInfo handle name dir st end ascii watchedRegs
   when (shouldBreak dir breakPoint) waitForUser
 
-printInfo :: Handle -> String -> Direction -> Internal.State -> String -> Bool -> [WatchedReg] -> Internal.RT ()
+printInfo :: Handle -> String -> Direction -> Internal.State -> String -> Bool -> [WatchedReg] -> RT ()
 printInfo handle name dir st@Internal.State{input, line, col} end ascii regs = do
   let cs = replace "\n" (newline ascii)
          . replace " " (space ascii)
@@ -146,12 +146,12 @@ printInfo handle name dir st@Internal.State{input, line, col} end ascii regs = d
     if null regs then return []
     else (++ [""]) . ("watched registers:" :) <$>
       forM regs (\(WatchedReg rname reg) ->
-        (\x -> "    " ++ rname ++ " = " ++ show x) <$> readReg reg)
-  Internal.unsafeIOToRT $
+        (\x -> "    " ++ rname ++ " = " ++ show x) <$> readRef reg)
+  unsafeIOToRT $
     hPutStr handle $ indentAndUnlines st ((prelude ++ cs' ++ end) : caret : regSummary)
 
-waitForUser :: Internal.RT ()
-waitForUser = Internal.unsafeIOToRT $ do
+waitForUser :: RT ()
+waitForUser = unsafeIOToRT $ do
   echo <- hGetEcho stdin
   hSetEcho stdin False
   putStrLn "..."