+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   with pattern synonyms and `forall`s in more places.
 * Added line numbering for the line information in the `ErrorBuilder`.
 * Renamed the `format` method of the `ErrorBuilder` to `build`.
+* Removed `RT` from the library's internals, it is now part of the `rt` package.
 
 ## 0.2.5.1 -- 2024-02-07
 * Fixed bug where hints can be revived by the `hide` combinator.